+++ crates/cli/src/commands/mod.rs
@@ -9,7 +9,6 @@ mod install_global;
 mod list;
 mod list_global;
 mod list_remote;
-mod local;
 mod pin;
 mod plugins;
 mod remove_plugin;
@@ -32,7 +31,6 @@ pub use install_global::*;
 pub use list::*;
 pub use list_global::*;
 pub use list_remote::*;
-pub use local::*;
 pub use pin::*;
 pub use plugins::*;
 pub use remove_plugin::*;
