+++ crates/cli/src/main.rs
@@ -51,7 +51,6 @@ async fn main() -> MainResult {
         Commands::List(args) => app.execute_with_args(commands::list, args),
         Commands::ListGlobal(args) => app.execute_with_args(commands::list_global, args),
         Commands::ListRemote(args) => app.execute_with_args(commands::list_remote, args),
-        Commands::Local(args) => app.execute_with_args(commands::local, args),
         Commands::Pin(args) => app.execute_with_args(commands::pin, args),
         Commands::Plugins(args) => app.execute_with_args(commands::plugins, args),
         Commands::RemovePlugin(args) => app.execute_with_args(commands::remove_plugin, args),
